Reorganised UI and input labels added
* A set of labels added to show the user their inputs, this was added as
there was a space in the bottom right of the UI which needed filling.

* Labels added showing the user changings in the Cash Rate and other
data from the earliest input.

*Graphs updated; a click function was going to be added however the
application cannot read more than 3 windows forms possibly due to visual
studio addition but couldn't find enough details to fix the issue
unfortunately. The function showed a scalable graph on a new popup form
so the user could easily see the data on smaller screen etc.